pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "eml"
version = "0.1.0"
description = "Exp-Minus-Log (EML) operator as a softmax replacement in transformer attention."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "numpy>=1.24",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
where = ["src"]
